made welcome default, fixed sidebar link style (#378)
